rgb_panel: Add color mapper module to convert from color bits to planes
The fact that the color mapper can be multi-cycle makes the whole
readout logic more complex since we need a pipeline that needs to
stall. And this is also complicated by the fact that the SPRAM doesn't
support a READ_ENABLE and so when we stall we already issued the read
for the next address.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>